refactor: 미니게임 UI 패키지 구조 변경 및 관련 클래스 이동
- `ui` 패키지 내 코드를 `command`, `request`, `response` 등 세부 패키지로 분리
- 불필요 클래스(`CardGameSelectMessage`) 제거
