--- backend/src/main/java/coffeeshout/minigame/ui/CardGameSelectMessage.java
@@ -1,4 +1,0 @@
-package coffeeshout.minigame.ui;
-
-public record CardGameSelectMessage(String playerName, Integer cardIndex) {
-}
