+++ backend/src/main/java/coffeeshout/minigame/ui/command/MiniGameCommandHandler.java
@@ -1,4 +1,4 @@
-package coffeeshout.minigame.ui;
+package coffeeshout.minigame.ui.command;
 
 public interface MiniGameCommandHandler<T extends MiniGameCommand> {
