+++ backend/src/main/java/coffeeshout/minigame/ui/request/command/SelectCardCommand.java
@@ -1,7 +1,7 @@
-package coffeeshout.minigame.ui.handler;
+package coffeeshout.minigame.ui.request.command;
 
 import coffeeshout.minigame.domain.MiniGameType;
-import coffeeshout.minigame.ui.MiniGameCommand;
+import coffeeshout.minigame.ui.command.MiniGameCommand;
 
 public record SelectCardCommand(String playerName, Integer cardIndex) implements MiniGameCommand {
