+++ backend/src/test/java/coffeeshout/minigame/ui/CardGameIntegrationTest.java
@@ -10,8 +10,9 @@ import coffeeshout.minigame.application.CardGameService;
 import coffeeshout.minigame.domain.MiniGameType;
 import coffeeshout.minigame.domain.cardgame.CardGameRound;
 import coffeeshout.minigame.domain.cardgame.CardGameState;
-import coffeeshout.minigame.ui.MiniGameStateMessage.CardInfoMessage;
-import coffeeshout.minigame.ui.handler.SelectCardCommand;
+import coffeeshout.minigame.ui.response.MiniGameStateMessage;
+import coffeeshout.minigame.ui.response.MiniGameStateMessage.CardInfoMessage;
+import coffeeshout.minigame.ui.request.command.SelectCardCommand;
 import coffeeshout.minigame.ui.request.CommandType;
 import coffeeshout.minigame.ui.request.MiniGameMessage;
 import coffeeshout.room.domain.JoinCode;
